Dealing with warnings in Windows builds (except number 9: regression issue in EigenDecomposition::checkEigenValueEquation). 1. In alignment.cpp, lots of references to cout qualified (to std::cout) 2. Simplified a string-comparison expression in     Alignment::readCountsFrom() (Visual Studio thought that     || operators should've been | operators, because it was relying     on implicit conversion to compare int values from string::compare     with zero). 3. Alignment::shouldRemoveSequence(), changed the seq1 and seq2     parameters (they were int, should've been intptr_t). Likewise,     (a) Alignment::areSequencesIdentical()     (b) Alignment::reportSequenceKept() 4. In SuperAlignment::checkAbsentStates(), index should be int64_t. 5. In ModelFileLoader::parseModelNameAndParameters(),     the param_start variable is now a size_t (it should always have been) 6. In model/modelmarkov.cpp, rescaleMatrix()'s 4th parameter     (total_num_subst) should be declared as a double (it now is). 7. In ModelPoMoMixture, the (visual studio only) declares of the     virtual function implementations that were being mapped to     ModelMixture::X (for whatever member function X) have to be     implemented via calls to mixture.X (mixture is an instance of     ModelMixture owned by ModelPoMoMixture). 8. Added an int cast in IQTree::deleteNonCherryLeaves     and a couple more in IQTree::initSettings. 9. Fixed up a bug in EigenDecomposition::checkEigenValueEquation     (added when I created it: zero needs to be declared double).     Also renamed error (to is_error) and made it a bool, not an int,     and rigged the outer (j) for-loop to exit early if is_error goes true. 10. PhyloTree::constructSimilarityTree(), interior_node_id needs to     be an intptr_t rather than an int. 11. ParsimonyRouter::constructTree(), stop_batch likewise.       And start_batch, and i, and j. 12. PhyloTree::mergeAlignment(), merged_seq_id likewise (twice). 13. Broke up some strings that were too long for Visual Studio,      in modelcodonempirical.cpp 14. In whtest/weisslambda_sub.c's ComputeWeissLambdaQ16,       paare has to be an intptr_t (otherwise multiplication may       overflow). And i must be too, otherwise for-loop might run       forever (i can't reach paare).

+++ alignment/alignment.h
@@ -546,12 +546,12 @@ public:
                                           StrVector &target_seqs);
 
         //Supporting functions for removeIdenticalSeq...
-        bool shouldRemoveSequence(int seq1, int seq2, 
-                                const string& not_remove,
-                                const BoolVector& isSequenceRemoved,
-                                const vector<size_t>& hashes) const;
+        bool shouldRemoveSequence(intptr_t seq1, intptr_t seq2, 
+                                  const string& not_remove ,
+                                  const BoolVector& isSequenceRemoved,
+                                  const vector<size_t>& hashes) const;
 
-        void reportSequenceKept(int seq1, int seq2, bool listIdentical,
+        void reportSequenceKept(intptr_t seq1, intptr_t seq2, bool listIdentical,
                                 progress_display& progress) const;
 
         void doneCheckingForDuplicateSequences(double startCheck,
